remote/client: Ensure driver is created only for specified resource
When using `power` command, one can use `-n` to choose to which resource
on a target the command will apply. However, when looping over resources
to find a compatible driver, it ignores the named resouce. This will
cause an exception if another resource is chosen due another driver
being selected on the loop.

This patch fix it by ignoring resources whose name do not match the
resource given in the command line.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>